js/xpconnect/tests/mochitest/test_bug781476.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=781476
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 781476</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=781476">Mozilla Bug 781476</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 781476 **/
SimpleTest.waitForExplicitFinish();

function go() {
  var iwin = document.getElementById('ifr').contentWindow;
  iwin.is = is;
  var evt = iwin.makeEvent();
  is(evt.expando, 42, "Expando properly visible in caller frame");
  SimpleTest.finish();
}


</script>
</pre>
<iframe onload="go();" id="ifr" src="file_bug781476.html"></iframe>
</body>
</html>
